Update Get-SCEPCertificate.ps1
Modifies Get-SCEPCertificate.ps1 to adjust the behavior of the -Suffix parameter. The previous behavior forced the configuration string to always end with "/pkiclient.exe" which is a standard endpoint but is not always implemented appropriately by certificate issuers - as an example, when performing SCEP enrollment with Okta as the Certificate Authority, their certificate endpoints never end with this suffix.

Now, the -Suffix parameter may optionally specified as an empty string, which will result in nothing being appended to the configuration string after the port number. Situations where -Suffix is not specified have identical behavior as the "/pkiclient.exe" suffix is part of the default parameter value.